Former-commit-id: 9c9a3c27d38676b2c992aca129185df9ca2631f1
Former-commit-id: cd86254eda56f003bd2150320cf1ed3df004e9bb

+++ backend/src/main/java/com/ssafy/i5i/hotelAPI/domain/food/controller/FoodController.java
@@ -20,7 +20,7 @@ public class FoodController {
 	@GetMapping("/by-name")
 	public List<FoodResponseDto.Coordi> getFoodFromTravle(
 			@RequestParam("name") String name,
-			@RequestParam("distance") Long distance,
+			@RequestParam("distance") Double distance,
 			@RequestParam(value = "sorted", required = false) String sorted){
 		FoodRequestDto.Title attractionTitleRequestDto = new FoodRequestDto.Title(name, distance, sorted);
 		return foodServiceImpl.getFoodFromTravle(attractionTitleRequestDto);
@@ -29,7 +29,7 @@ public class FoodController {
 	public List<FoodResponseDto.Coordi> getFoodFromLngLatv(
 			@RequestParam("longitude") Double longitude,
 			@RequestParam("latitude") Double latitude,
-			@RequestParam("distance") Long distance,
+			@RequestParam("distance") Double distance,
 			@RequestParam("sorted") String sorted){
 		FoodRequestDto.Coordi attractionCoordiRequestDto = new FoodRequestDto.Coordi(latitude, longitude, distance, sorted);
 		return foodServiceImpl.getFoodFromLngLatv(attractionCoordiRequestDto);
